:sparkles: feat: finalize header component
